scrapers/requirements.txt
playwright==1.49.0
feedparser==6.0.11
anthropic>=0.40.0
requests>=2.31.0
python-dotenv>=1.0.0
